Adding pytest to requirements. Putting mock definitions in conftest
