+++ requirements.txt
@@ -6,6 +6,7 @@ redis==3.5.3
 redisgraph==2.1.5
 redisgraph-bulk-loader==0.9.5
 requests<2.24.0
+pytest==6.2.2
 PyYAML==5.3.1
 git+git://github.com/biolink/biolink-model-toolkit@0.4.0#egg=bmt
 git+git://github.com/helxplatform/dug@2.0.1#egg=dug
